requirements.txt
flatbuffers==23.5.26
faiss-cpu==1.7.4
duckdb==0.9.2
numpy==1.26.0
scipy==1.11.3
sentence-transformers==5.1.2
open_clip_torch==3.2.0
torch
pillow
nltk==3.8.1
tqdm
lmdb
wikipedia
ddgs
trafilatura